Fix race condition in the watch feature
This affects issue #10, possibly resolving it.

This commit fixes a race condition due to the watch object being destructed and
reconstructed after each file system event. It's possible for a file system
event to occur after one of these destructions but before the next
construction such that a second, subsequent file system event is lost.

Here's a possible scenario:

1. File system event A occurs.
2. The watch object destructs.
3. The template engine is reloaded.
4. File system event B occurs.
5. A new watch object is constructed.

In this scenario, the second watcher will not detect event B. If B represents a
template being updated then Handlebars Iron will not pick up the change.